fix(frontends/lean,shell): fix file dependencies
