+++ src/frontends/lean/CMakeLists.txt
@@ -8,4 +8,4 @@ init_module.cpp type_util.cpp decl_attributes.cpp
 prenum.cpp print_cmd.cpp elaborator.cpp
 match_expr.cpp local_context_adapter.cpp decl_util.cpp definition_cmds.cpp
 brackets.cpp tactic_notation.cpp info_manager.cpp json.cpp tactic_evaluator.cpp
-equations_validator.cpp parser_state.cpp interactive.cpp)
+equations_validator.cpp parser_state.cpp interactive.cpp completion.cpp)
