+++ src/frontends/lean/interactive.cpp
@@ -11,10 +11,10 @@ Author: Sebastian Ullrich
 #include "library/module_mgr.h"
 #include "library/versioned_msg_buf.h"
 #include "library/attribute_manager.h"
+#include "frontends/lean/completion.h"
 #include "frontends/lean/interactive.h"
 #include "frontends/lean/pp.h"
 #include "frontends/lean/tactic_notation.h"
-#include "shell/completion.h"
 
 namespace lean {
 LEAN_THREAD_VALUE(break_at_pos_exception::token_context, g_context, break_at_pos_exception::token_context::none);
